osd/scrub: do not reduce min chunk on preemption
Existing logic handling scrub preemptions is halving both the
'max' and the 'min' values.
This isn't optimal: the 'min' values (used mainly to guarantee a minimal
number of objects to fetch from the backend in one operation) can and
should also be used to limit the effect of preemptions on the execution
of the scrub.

Fixes: https://tracker.ceph.com/issues/73410
Signed-off-by: Ronen Friedman <rfriedma@redhat.com>
(cherry picked from commit 3e8ec7d08ac7a7302abd27185dca2660df9c429f)