src/mon: Add validation for the w value when creating a blaum-roth EC profile to OSDMonitor.cc. src/erasure-code/jerasure: remove the validation for blaum-roth w value from ErasureCodeJerasure.cc.
This commit removes the validation for the blaum-roth w value from ErasureCodeJerasure.cc and moves it to OSDMonitor.cc. This allowed me to also add a condition to bypass the validation using the flag --yes-i-really-mean-it.
This commit also adds an is_prime() method to OSDMonitor.cc which is used in the w value validation.
This commit also fixes an error in HealthMonitor.h where I missed a semi-colon on the end of a line.

Fixes: http://tracker.ceph.com/issues/64419
Signed-off-by: Tom Sollers <tom.sollers@ibm.com>