mgr/dashboard: Fix dashboard api tests in teuthology
Fixes https://tracker.ceph.com/issues/70669

This issue happens when iostat module enabled -> mgr respawned -> dashboard restart (attempts to bind to 8443) -> Port 8443 unavailable -> and API cant access.
This is a race condition between process restart and OS port release. It is worsened by rapid restarts, slow cleanup on shared/flaky test hosts like Teuthology,
It's not an inherent server bug or misbehavior in the dashboard service. This failure is a test/lifecycle/environment flake.

We need to find a suitable time window here, probably increasing the time to wait for mgr re-initialization and port to be assigned to dashboard.

Increasing timeouts and removing nested retry mechanisms to properly debug and trace tests.

Signed-off-by: Afreen Misbah <afreen@ibm.com>