cephadm/smb: Publish SMB port from config than constant
This help us switch the default listening ports (445 and 139) for smbd
to the configured custom port or 445 without host networking enabled.
It is done for consistency among clustered and non-clustered modes to
always and only publish the listening smbd port/s. Previously without
host networking the legacy port 139 was never published even though it
was listening inside the container.

Signed-off-by: Anoop C S <anoopcs@cryptolab.net>