mgr/smb: Always use 'smb ports' parameter for smbd
Irrespective of whether custom port is configured for smb, use the
'smb ports' parameter in smb.conf. This would mean that we only listen
on 445 and not the legacy port 139 unless explicitly configured. With
smb daemon already modified to publish whatever is the listening port
from config without host networking, it makes sense to get rid of the
corresponding check while adding the 'smb ports' parameter.

Signed-off-by: Anoop C S <anoopcs@cryptolab.net>