qa/standalone/erasure-code: Add a new test to test-erasure-code-plugins.sh to test for new health warn
This commit adds a new test to test-erasure-code-plugins.sh that tests for the health warning caused by having a erasure-code-profile with the blaum-roth technique and a w+1 value that is not prime.
This commit also fixes the health warning to include the w+1 value for the profile whos w+1 is not prime.

Fixes: http://tracker.ceph.com/issues/64419
Signed-off-by: Tom Sollers <tom.sollers@ibm.com>