osd: pool migration skeleton peering state machine
Skeleton changes to peering state machine for pool migration.

Add a new state WaitLocalPoolMigration that is called after recovery
and backfill have completed if a pool migration is required.

Reserve local resources for a pool migration and transition to state
WaitRemotePoolMigration.

WaitRemotePoolMigration is a stub state that releases the local
reservations and transitions to Recovered. If the local reservation
is cancelled a DeferPoolMigration event is generated and currently
ignored.

Signed-off-by: Bill Scales <bill_scales@uk.ibm.com>