/src: Add back the validation to ErasureCodeJerasure.cc
In this commit i have added back the validation to ErasureCodeProfile.cc as it could be called by other areas of the codebase without going through the OSDMonitor.cc. Changed the test in test-erasure-code-plugins.sh to reflect the new health warning, there is no need for two tests as the output is the same regardless of the number of bad profiles.

Fixes: http://tracker.ceph.com/issues/64419

Signed-off-by: Tom Sollers <tom.sollers@ibm.com>