messages: fix MOSDOpReplay decode of old version bad_replay_version
When decoding an old version of MOSDOpReplay or when the PGID64
feature is enabled, bad_replay_version is not being set, which
causes inconsistent output in ceph-dencoder.

To fix this, set bad_replay_version to the head reassert_version
value. This ensures that if the struct is later re-encoded, the
decode output will be consistent.

Fixes: https://tracker.ceph.com/issues/69814
Signed-off-by: Nitzan Mordechai <nmordech@ibm.com>