test/librados: Extend librados tests to cover both fast EC and split ops.
The librados tests have special handling for EC, but they do not attempt to test
either fast EC or split ops. This change upgrades the tests to be paramaterized
tests and constructs the necessary pools and boilerplate.

There is also a minor tweak to a stat tests to allow for longer names which
get generated in my containerised test environment.

Signed-off-by: Alex Ainscow <aainscow@uk.ibm.com>