mgr/cephadm: fixed bootstrap version adding process
I removed the add bootstrap code from both the upgrade process
and the get cluster version history command. These are almost
completely redundant as the add bootstrap code exist in the
Cephadm serve loop which loops every minute. This has the
effect of reducing confusion in the code flow and removes the
issue of duplicate entries being added for clusters that are
upgrading from a version without the feature, to a version with
the feature. I've also added code to ensure that the variable
responsible for checking if bootstrap has been added is set
properly, since this variable is reset when a cluster is
upgraded. Minor documentation update as well.

Signed-off-by: Timothy Q Nguyen <timqn22@gmail.com>