mgr/status: handle case where osd metadata is unavailable
The status module currently assumes that OSD metadata is always
available if the OSD exists in osd_stats (implied by an assertion
after calling metadata). However, in some cases,
(i.e. when a host is drained), the metadata becomes unavailable,
and a None type is returned when we try to retrieve it.

Since we only use the metadata to collect the hostname for each
OSD in osd_stats, we can skip collecting it when the metadata is
unavailable and simply print an empty string.

Fixes: https://tracker.ceph.com/issues/70164
Signed-off-by: Laura Flores <lflores@ibm.com>