mds: session reclaim could miss blocklisting an old session
Credit to ymo zhang[0] for the RCA and the fix.

This is due to a race between a client session close (which is
retired by the MDS) and a session open sent by the client due
to MDSMap update for the MDS state change. This results in the
old session still holding the client uuid (in client metadata).

[0]: https://tracker.ceph.com/issues/66003#note-8

Fixes: https://tracker.ceph.com/issues/66003
Signed-off-by: Venky Shankar <vshankar@redhat.com>
(cherry picked from commit c3ee87cff64f8222a57c23cfbb59b9add3b9677d)