script/build-with-container: fix passing branch names on cli
Previously, we converted slashes to dashes in a branch name when
determining the branch name via git but not if the branch name
was passed in on the cli. Since we're using BWC in automation
we can't expect the branch name to be always clean to directly
use in a container name (as we've seen) - so we will now
always convert slashes to dashes in a branch name no matter
how it was gotten.

Signed-off-by: John Mulligan <jmulligan@redhat.com>