qa/cephfs: cleanup test_rmtree_aborts_when_should_cancel_is_true
Remove "cancel_flag.clear()" as it can confuse reader that clearing the
event has an effect on rmtree() (but the fact is it doesn't). Removing
it is safe since it has zero impact impact.

Also a add a little sleep and a comment regarding this sleep in the same
test method.

Signed-off-by: Rishabh Dave <ridave@redhat.com>