rgw/logging: make sure source bucket is in the target's list
if a target log bucket is deleted, and then recreated with the same name,
the attribute holding the sources will not exist.
this is fixing it, but updating the list anytime we try to flush
or add a record.
if the source is already in the list, there will be no update to the attribute.

Fixes: https://tracker.ceph.com/issues/73675

Signed-off-by: Yuval Lifshitz <ylifshit@ibm.com>