libcephfs_proxy: add the number of supported operations to negotiation
The v0 negotiation structure has been modified to hold the total number
of operations and callbacks supported by the peer. The changes are done in
a way that it's completely transparent and harmless for a peer expecting
the previous definition.

This will be useful to quickly check if the daemon supports some
operation, or the client supports some callback before sending them.

Signed-off-by: Xavi Hernandez <xhernandez@gmail.com>