osd: Tidy up RepNotRecovering in peering state machine diagram
RepNotRecovering state has to deal with race conditions in
recovery/backfill/pool migration where either a local event
or a RELEASE message from the primary can stop the background
process and if both happen near simultaneously this state has to
ignore the second event.

To avoid clutter on the peering state machine diagram in the
documentation these uninteresting events can be coded as
reactions that discard the event (which are not shown on the
diagram) rather than as trivial transitions to the same state.

Signed-off-by: Bill Scales <bill_scales@uk.ibm.com>