rados: Add API to disable version querying with reads in librados
librados will always request a "user version". Until EC direct reads are implemented
this is a cheap operation and so librados always requests the user version, even if
the client does not need it.

With EC direct reads, requesting the user version requires an extra op to the primary
in some scenarios. The non-primary OSDs do not contain an up to date user
version.

NEORADOS already allows for such optimisations, due to a how the API is organised.

librados is not heavily used by ceph-maintained clients, but this API will still be
useful for testing of EC direct reads, since the test clients will use librados, due
to it simpler nature and performance not being critical in the tests.

Signed-off-by: Alex Ainscow <aainscow@uk.ibm.com>