cephadm: fix zip_root_entries population in version command
The 'cephadm version --verbose' command was returning an empty
zip_root_entries list because it relied on the private '_files'
attribute of zipimport.zipimporter, which is not reliably populated
across Python versions.

This commit fixes the issue by using the zipfile module to properly
read the archive contents via the loader.archive path. This ensures
that zip_root_entries is correctly populated with the root-level
directories in the zipapp.

This fix is necessary for the cephadm build tests to properly validate
that all expected packages and modules are included in the built zipapp.

Signed-off-by: Kefu Chai <k.chai@proxmox.com>
(cherry picked from commit 2c68c1496dbb7cd01bf783e31510940445040a34)