rbd-mirror: handle ERESTART errno in GroupReplayer handle_shut_down()
if the start of the GroupReplayer falls behind the group disable/remove
CLI command,  the GroupReplayer bootstrap logic attempting to
PrepareLocalGroupRequest sees the group is in disabling state and sends
a ERESTART to GroupReplayer.

But, currently in the GroupReplayer set_finished() is never called in case if
the error is ERESTART. This blocks the group from not getting removed forever
as the InstanceReplayer::start_group_replayer() looks at is_finished() only
then attempts to destroy the GroupReplayer object from m_group_replayers map.

Signed-off-by: Prasanna Kumar Kalever <prasanna.kalever@redhat.com>