osd: Fail I/O to objects that have been migrated
Fail attempt to do I/O to objects that have been
migrated from the source to target pool. Eventually
the client will check for this new failure code and
retry the I/O to the target pool. The client will
lazily cache the watermark tracking the progress of
the migration so will occasionally send requests
to the wrong pool.

Signed-off-by: Bill Scales <bill_scales@uk.ibm.com>