src/script: add script to store pool availability status
This commit adds a script, pool-status-check.sh, to enable
testing of the data availability feature. The script specifically
stores the ceph status output and pool's data availability
output in a log file every 10 mins to a log file. This log file
can be used to analyse the reliability of the feature. The script
ensures multiple instances of the same script are not run.

Command to run the script such that the terminal is not
blocked and losing SSH connection will not interrupt the
script:

	nohup ./pool_status_check.sh >/dev/null 2>&1 &

Command to stop the script:

	./pool_status_check.sh --stop

Signed-off-by: Shraddha Agrawal <shraddhaag@ibm.com>