osd: Calculate high watermark for pool migrations
Pool migration requires a high watermark which tracks which
objects in a PG have been migrated to the target pool. Unlike
backfill which persists this in pg_info_t; migration can calcuate
the watermark by scanning the object store for the earliest
hobject because objects are deleted as they are migrated.

Whilst pool migration normally only occurs while the PG is
active+clean it is possible that the process is interrupted
by a backfill/recovery and in these cases the watermark must
persist even though there may be missing objects in the
object store. Therefore like PG_NLS the watermark is
determined by looking at both the object store and the
missing list (including missing deletes).

Signed-off-by: Bill Scales <bill_scales@uk.ibm.com>