qa: Improve failure reporting for watchdog timeouts
A daemon such as an osd or mon crashing can lead to teuthology tests
failing, these failures are detected by the daemon watchdog. Currently
the failure message just indicates that a watchdog timeout occurred.

Add an AssertTracker class that processes the stderr output from
daemon processes, taking insipration from teuthology scrape.py to
capture the assert string and backtrace if the daemon crashes

Enhance the watchdog to report the assertion and backtrace if
a daemon has crashed so this gets reported as the teuthoolgy
job failure reason

Add WatchedProcess for radosbench

Signed-off-by: Bill Scales <bill_scales@uk.ibm.com>