osd: EC changes to advance pool migration watermark as objects are deleted
Changes to EC (both fast EC and old EC) to call out to PrimaryLogPG
when objects are deleted. This can reply with an updated migration
watermark which is then sent in the ECSubOpWrite message and passed
to PrimaryLogPG on all shards as the transaction is applied to update
the watermark.

PrimaryLogPG is modified so that PGs do not signal the end of a
migration until the watermark has advanced to MAX.

Code in PrimaryLogPG is a temporary low performance implementation
and will be replaced in a later commit.

Signed-off-by: Bill Scales <bill_scales@uk.ibm.com>